requirements.txt
Django==1.7.1
djangorestframework==3.0.2
django-cors-headers==1.0.0
pytz==2014.10
django-extensions==1.4.9
mock==1.0.1
requests==2.5.1
psycopg2==2.5.4
numpy==1.9.2
